requirements.txt
torch
transformers
datasets[audio]
huggingface_hub
pyyaml
librosa
soundfile
numpy
python-dotenv
omegaconf
jiwer
peft
accelerate
torchaudio